+++ CHANGELOG.md
@@ -1,8 +1,23 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.0.2] - 
  
-## [0.0.1] - 2024-09-26
+### Added
+
+
+### Changed
+
+- Added ssh connect timeout of 5s
+ 
+### Fixed
+
+- Small fix in the Shell.tsx
+- Centered loading screen
+- Handling of uncaught exceptions
+ 
+ 
+## [1.0.1] - 2024-09-26
  
 ### Added